Remove unneeded command, rename indexing ones
