+++ search/management/commands/reindex_search_engine_forum.py
